% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclim_hydrotherm.R
\name{bioclim_hydrotherm}
\alias{bioclim_hydrotherm}
\title{Calculation of hydrothermal viticultural indices (Branas, Dryness Index)}
\usage{
bioclim_hydrotherm(climdata, lat, elev)
}
\arguments{
\item{climdata}{a dataframe with daily weather data.
Must contain the columns Year, Month, Day, Tmax, Tmin, RHmax, RHmin, Prec, 
Rad, u2.}

\item{lat}{the latitude of the site, in decimal degrees. Use positive values
for Northern latitudes and negatives for Southern.}

\item{elev}{the elevation of the site, in meters above sea level.}
}
\value{
data frame with the values of the indices for each season in the
climdata file.
}
\description{
This function calculates the hydrotermic index of Branas, Bernon and 
Levandoux (Branas et al 1946) and the Dryness index (Riou et al 1994).
The Branas Index (BBLI) takes  into  account  the influence  of  both  
temperature  and  precipitation  on  grape yield  and  wine  quality.  
This  index  is  the  sum of  the products of monthly mean temperature 
(Tmean,in Celsius) and monthly  accumulated  precipitation  amount  (Prec,in mm)
during the April to September season (Northern Hemisphere) or October to 
February (Southern Hemisphere).
}
\details{
The  dryness  index  (DI)  is  measured  based  on  an adaptation  of  
the  potential  water  balance  of  the  soil index of Riou (Riou et al., 
1994), developed specially for vineyard use. It  enables  the  characterization
of the  water  component  of  the  climate  in  a  grape-growing  region,
taking into account the climatic demand of a standard vineyard, evaporation
from bare soil, rainfall without deduction for surface runoff or drainage. 
It indicates the  potential  water  availability  in  the  soil,  related  
to the level of dryness in a region (Tonietto and Carbonneau, 2004). The
index uses potential evapotranspiration using the Penman Monteith method.
}
\examples{

\dontrun{

#select the appropiate columns from a larger dataset with date information
#in Year, Month, Day format, create a vector or harvest dates and
#estimate the Huglin and Cool night indices on each year
#in the series.

Weather <- Tempdata \%>\%
   select(Year, Month, Day, Tmax, Tmin)
harvest <- c(225, 250, 275)
latitude <- 42.08
Vitic_indices <- viticultural_indices(Weather, harvest, latitude)

}
}
\references{
Riou C, Carbonneau A, Becker N, Caló A, Costacurta A, Castro R, Pinto PA, 
Carneiro LC, Lopes C, Climaco P, Panagiotou MM, Sotes V,Beaumond HC, Burril A, 
Maes J, Vossen P. 1994.Le determinisme climatique de la maturation du raisin: 
application au zonage de la teneur em sucre dans la communaute europenne. 
Office des Publications Officielles des Communautes Europennes: Luxembourg, 322pp.

Tonietto J, Carbonneau A. 2004. A multicriteria climatic classification system 
for grape-growing regions worldwide. Agricultural and Forest Meteorology, 124:81-97.
}
\author{
Carlos Miranda
}
