% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclim_thermal.R
\name{bioclim_thermal}
\alias{bioclim_thermal}
\title{Calculation of thermal viticultural indices (Huglin, Winkler and Cool Nights)}
\usage{
bioclim_thermal(climdata, lat)
}
\arguments{
\item{climdata}{a dataframe with daily maximum and minimum temperatures.
Must contain the columns Year, Month, Day, Tmax, Tmin.}

\item{lat}{the latitude of the site, in decimal degrees. Use positive values
for Northern latitudes and negatives for Southern.}
}
\value{
data frame with the values of the indices.
}
\description{
This function calculates the Heliothermal Index of Huglin (1978),
the Winkler index (Amerine and Winkler, 1974) and the Cool Night 
index (Tonietto, 1999).
}
\examples{

\dontrun{

#select the appropiate columns from a larger dataset with date information
#in Year, Month, Day format, create a vector or harvest dates and
#estimate the Huglin and Cool night indices on each year
#in the series.

Weather <- Tempdata \%>\%
   select(Year, Month, Day, Tmax, Tmin)
harvest <- c(225, 250, 275)
latitude <- 42.08
Vitic_indices <- bioclim_thermal(Weather, harvest, latitude)

}
}
\references{
Amerine MA and Winkler AJ. 1944. Composition and quality of musts and wines
of California grapes. Hilgardia 15: 493-675.

Huglin P. 1978. Noveau mode d'evaluation des possibilites héliothermiques
d'un milieu viticole. In: Proceedings of the Symposium International sur
l'ecologie de la Vigne. Ministére de l'Agriculture et de l'Industrie 
Alimentaire, Contança pp 89-98.

Tonietto J. 1999. Les macroclimats viticoles mondiaux et l'influence du
mésoclimat sur la typicité de la Syrah et du Muscat de Hambourg dans le
sud de la France: methodologie de carácterisation. Thése Doctorat. Ecole 
Nationale Supérieure Agronomique, Montpellier, 233pp.
}
\author{
Carlos Miranda
}
