% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_potential.R
\name{color_potential}
\alias{color_potential}
\title{Evaluation of weather conditions for anthocyain formation
in apple skin}
\usage{
color_potential(climdata, harvest)
}
\arguments{
\item{climdata}{a dataframe with daily maximum and minimum temperatures.
Must contain the columns Year, Month, Day, Tmax, Tmin.}

\item{harvest}{a vector with expected harvest days
(expressed as day of the year)}
}
\value{
data frame with the number of favorable (Cool_d) and unfavorable
(Hot_d) days for red color, for each year (Year) in the series and
harvest date (Day_h) provided.
}
\description{
This function estimates the number of days favorable and
unfavorable for anthocyanin accumulation in the skin of
red apple cultivars during the month (30 days) before harvest.
A favorable day (Cool day) is considered when the daily maximum
temperature is below 26ºC, an unfavorable day (Hot day) when
the minimum temperature is above 20ºC (Lin-Wang et al, 2011). The
function allows testing several harvest dates, supplied as
a vector.
}
\examples{

\dontrun{

#select the appropiate columns from a larger dataset with date information
#in Year, Month, Day format, create a vector or harvest dates and
#estimate the number favorable and unfavorable days on each year
#in the series.

Weather <- Tempdata \%>\%
   select(Year, Month, Day, Tmax, Tmin)
harvest <- c(225, 250, 275)
Colorassess <- color_potential(Weather, harvest)

}
}
\references{
Lin-Wang K, Micheletti D et al, 2011. High temperature reduces apple fruit
colour via modulation of the anthocyanin regulatory complex. Plant, Cell
and Environment 34, 1176-1190.
}
\author{
Carlos Miranda
}
