% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_frost.R
\name{days_frost}
\alias{days_frost}
\title{Estimates the occurence of spring frosts}
\usage{
days_frost(mintemps, fendates, lastday = 181)
}
\arguments{
\item{mintemps}{a dataframe with minimum daily temperatures. Must contain
the columns julian day of year (DOY )and the minimum daily temperature (Tmin).}

\item{fendates}{a dataframe with the day of occurrence of the phenological
stages. Must contain julian day of occurrence of the phenological
stage (DOY) and the critical frost temperature for the stage (Tcrit)}

\item{lastday}{the last day (day of the year) to evaluate. By default, 
lastday = 181 (June 3oth).}
}
\value{
a dataframe with the columns
number of days in which Tmin is equal or below Tcrit.
}
\description{
The function evaluates the early and spring frosts occurring
during a season. It compares minimum daily temperature (Tmin) with the critical
temperature (Tcrit). Daily critical temperatures are linearly
interpolated from a user-provided dataframe with critical temperatures for
key phenological stages and the day of occurrence of the stages.
The function checks each day between the first date if Tmin is below Tcrit
and, if so, that day is considered as a frost day (Fday). The last day evaluated
is DOY 181.
}
\examples{

\dontrun{
#select the appropiate columns from a larger dataset with date information
#in Year, Month, Day format, include date and DOY information and estimate
#the #number of frost days for the year 2018

Tmins_2018 <- Tempdata \%>\%
   select(Year, Month, Day, Tmin) \%>\%
   mutate(Date=make_date(Year,Month,Day),DOY=yday(Date)) \%>\%
   filter(Year==2018)
Frostdays_2018 <-days_frost(Tmins_2018,Tcrit)
}
}
\author{
Carlos Miranda
}
\keyword{frost}
\keyword{prediction}
