% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_Penman.R
\name{ET_penman}
\alias{ET_penman}
\title{Calculation of dailiy potential evapotranspiration by Penman (1948) method}
\usage{
ET_penman(climdata, lat, elev)
}
\arguments{
\item{climdata}{a dataframe with daily weather data.
Must contain the columns Year, Month, Day, Tmax, Tmin, RHmax, RHmin, Rad, u2.}

\item{lat}{the latitude of the site, in decimal degrees.}

\item{elev}{the elevation of the site, in meters above sea level.}
}
\value{
climdata data frame in which Date, DOY and ET columns have been added.
}
\description{
This function calculates the potential evapotranspiration (ETref) using daily 
weather data and the Penman (1948) method
}
\details{
This version of the function requires the user to supply in weather data daily
values for temperature (Tmax and Tmin), relative humidity (RHmax and RHmin), 
solar radiation (Rad in MJ m-2 day-1) and wind speed at 2m height(u2 in m s-1).
}
\examples{

\dontrun{

elevation <- 315
latitude <- 42.08
ET_Penman <- ET_penman(Weather, latitude, elevation)
}
}
\references{
To be included
}
\author{
Carlos Miranda
}
