% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_Penman_Monteith.R
\name{ET_penman_monteith}
\alias{ET_penman_monteith}
\title{Calculation of dailiy reference evapotranspiration by Penman-Monteith method}
\usage{
ET_penman_monteith(climdata, lat, elev)
}
\arguments{
\item{climdata}{a dataframe with daily weather data.
Must contain the columns Year, Month, Day, Tmax, Tmin, RHmax, RHmin, Rad, u2.}

\item{lat}{the latitude of the site, in decimal degrees.}

\item{elev}{the elevation of the site, in meters above sea level.}
}
\value{
data frame with Year, Month, Day, DOY, ETos and ETrs values.
}
\description{
This function calculates the reference evapotranspiration (ETref) for short
(ETos) and tall (ETrs) canopies using daily weather data. The method is based
on the FAO56 guidelines (Allen et al, 1998) and on the standardized Penman Monteith 
equation from the Environmental Water Resources Institute of the American 
Society of Civil Engineers (Allen et al, 2005).
}
\details{
This version of the function requires the user to supply in weather data daily
values for temperature (Tmax and Tmin), relative humidity (RHmax and RHmin), 
solar radiation (Rad in MJ m-2 day-1) and wind speed at 2m height(u2 in m s-1).
}
\examples{

\dontrun{

elevation <- 315
latitude <- 42.08
ET_PM <- ET_penman_monteith(Weather, latitude, elevation)
}
}
\references{
Allen RG, Pereira LS, Raes D, Smith M. 1998. Crop evapotranspiration. Guidelines
for computing crop water requirements. FAO Irrigation and drainage paper 56. Food 
and Agriculture Organization of the United Nations

Allen RG, Walter IA, Elliott RL, Howell TA, Itenfisu D, Jensen ME, Snyder RL 2005. 
The ASCE standardized reference evapotranspiration equation. Reston, VA:American 
Society of Civil Engineers. 59 p.
}
\author{
Carlos Miranda
}
