% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDD_linear.R
\name{GDD_linear}
\alias{GDD_linear}
\title{Calculates growing degree days using linear method}
\usage{
GDD_linear(Temp_Day, Tb, Topt, Tcrit)
}
\arguments{
\item{Temp_Day}{a dataframe of daily temperatures. This data frame must
have a column for Year, Month, Day, and daily minimum ("Tmin") and
maximum ("Tmax") temperatures}

\item{Tb}{the base temperature to calculate GDD}

\item{Topt}{the optimal temperature to calculate GDD}

\item{Tcrit}{the critical temperature to calculate GDD}
}
\value{
data frame consisting of the columns Year, Month, Day, Tmax, Tmin,
Tmean and GDD.
}
\description{
The function calculates the daily heat unit accumulation (GDD)
from daily temperature data with a linear method based on averaging
the daily maximum and minimum temperatures (Arnold, 1960). Heat accumulation
begins when mean temperatures are above a minimum (base temperature,
Tb), and growth increases with linearly with temperature up to a
point (optimum temperature, Topt) at which there is no longer
an increase. The critical temperature (Tcrit) is the maximum
temperature at which growth will continue. The function allows
the user to define Tb, Topt and Tcrit.
}
\examples{

\dontrun{

GDH <- GDD_linear(Weather,4.5,25,36)

}

}
\references{
Arnold,  C.Y.  1960.  Maximum-minimum  temperatures as a basis for computing
heat units. Proc.Amer. Soc. Hort. Sci. 76:682–692.
}
\author{
Carlos Miranda
}
