% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_RH.R
\name{hourly_RH}
\alias{hourly_RH}
\title{Estimation of the hourly relative humidity on a daily series}
\usage{
hourly_RH(climdata, lat)
}
\arguments{
\item{climdata}{a dataframe with daily temperatures and RH for
each day in a series. Must contain the columns Year, Month, Day,
Tmax, Tmin, RHmax (daily maximum relative humidity) and RHmin
(minimum daily relative humidity).}

\item{lat}{latitude (decimal format) of the site}
}
\value{
data frame with columns Date, Year, Month, Day, DOY,
Hour, Temp and RH
}
\description{
This function estimates the hourly relative humidity (RH),
using daily temperature and humidity data. Hourly humidity is
estimated with the formula used by Waichler and Wigmosta (2003)
which require maximum and minimum values of daily temperature
and relative humidity. Hourly temperatures are calculated using
the function stack_hourly_temps from chillR package (Luedeling
et al, 2013).
}
\examples{

\dontrun{

#select the appropiate columns from a larger dataset with date information
#in Year, Month, Day format, include date and DOY information and estimate
#the hourly temperature and relative humidity for each day in the series.

Weather <- Tempdata \%>\%
   select(Year, Month, Day, Tmax, Tmin, WSmed) \%>\%
   mutate(Date=make_date(Year,Month,Day),DOY=yday(Date))
RH_h <- hourly_RH_temp(Weather, 41.5)

}
}
\references{
Luedeling E, Kunz A and Blanke M, 2013. Identification of chilling and heat
requirements of cherrytrees - a statistical approach.
International Journal of Biometeorology 57,679-689.

Waichler SR and Wigmosta MS, 2003. Development of hourly meteorological
values from daily data and significance to hydrological modeling at H.J.
Andrews experimental forest. Journal of Hydrometeorology 4, 251-263.
}
\author{
Carlos Miranda
}
