% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_windspeed.R
\name{hourly_windspeed}
\alias{hourly_windspeed}
\title{Estimation of the hourly wind speed from daily mean data}
\usage{
hourly_windspeed(climdata)
}
\arguments{
\item{climdata}{a dataframe with daily mean wind speed data (WSmed).
Must contain the columns Year, Month, Day and WSmed.}
}
\value{
data frame with the columns Date, Year, Month, Day, DOY,
Hour and WS (hourly wind speed).
}
\description{
This function estimates the hourly wind speed from a
dataset with mean daily wind speeds. Hourly wind speeds
from daily means are computed using a modified version of the
formula proposed by (Guo et al, 2016), so that maximum speeds
are obtained at mid-day.
}
\examples{
\dontrun{
Weather <- Tempdata \%>\%
   select(Year, Month, Day, WSmed) \%>\%
   mutate(Date=make_date(Year,Month,Day),DOY=yday(Date))
Windspeed_hourly <- hourly_windspeed(Weather)
}

}
\references{
Guo Z, Chang C, Wang R, 2016. A novel method to downscale daily wind
statistics to hourly wind data for wind erosion modelling. In: Bian F.,
Xie Y. (eds) Geo-Informatics in Resource Management and Sustainable
Ecosystem. GRMSE 2015. Communications in Computer and Information Science,
vol 569. Springer, Berlin, Heidelberg
}
\author{
Carlos Miranda
}
