% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderate_wind.R
\name{moderate_wind}
\alias{moderate_wind}
\title{Estimation of the daily hours with moderate wind from
daily weather data}
\usage{
moderate_wind(climdata)
}
\arguments{
\item{climdata}{a dataframe with daily mean wind speed data (WSmed).
Must contain the columns Year, Month, Day and WSmed.}
}
\value{
data frame with the columns Date, Year, Month, Day, DOY,
Hour and h_wind (hours with wind speed below 5.5 m/s).
}
\description{
This function estimates the daily hours with less than
'moderate breeze' wind (below 5.5 m/s in the Beaufort scale) from a
dataset with mean daily wind speeds. Hourly wind speeds
from daily means are computed using a modified version of the
formula proposed by (Guo et al, 2016), so that maximum speeds
are obtained at mid-day.
}
\examples{

\dontrun{

#select the appropiate columns from a larger dataset with date information
#in Year, Month, Day format, include date and DOY information and estimate
#the number favorable days on each year in the series

Weather <- Tempdata \%>\%
   select(Year, Month, Day, WSmed) \%>\%
   mutate(Date=make_date(Year,Month,Day),DOY=yday(Date))
hours_modwind <- moderate_wind(Weather)

}
}
\references{
Guo Z, Chang C, Wang R, 2016. A novel method to downscale daily wind
statistics to hourly wind data for wind erosion modelling. In: Bian F.,
Xie Y. (eds) Geo-Informatics in Resource Management and Sustainable
Ecosystem. GRMSE 2015. Communications in Computer and Information Science,
vol 569. Springer, Berlin, Heidelberg
}
\author{
Carlos Miranda
}
