% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_sequential.R
\name{phenology_sequential}
\alias{phenology_sequential}
\title{Prediction of phenological stages using a sequential model}
\usage{
phenology_sequential(GDH_day, Reqs, Start_chill)
}
\arguments{
\item{GDH_day}{a dataframe with daily chilling and forcing accumulation
must contain the columns Year, Month, Day, DOY, Chill, GD.}

\item{Reqs}{a dataframe with chilling and forcing requirements, must
contain the columns Creq (for chilling), Freq (for forcing heat).}

\item{Start_chill}{parameter indicating the day of the year when chill
accumulation is supposed to start.}
}
\value{
data frame with the predicted dates for every season in GDH_day
and stages in Reqs. For each season and phenological stage, the information
provided is the date in which chill requirements are satisfied (Year and
DOY), and the date in which the phenological stage is met (Year and DOY).
It contains the columns Creq, Freq, Season, Creq_Year,
Creq_DOY, Freq_Year, Freq_DOY.
}
\description{
The function predicts phenological phases for a climate series
from daily chill and heat requirements and daily chill and forcing data.
The sequential model used in the fucntion consider that chilling and heat
have independent effects. It consists of an accumulation of chill up to the
plant requirement followed by heat up to forcing requirement, with no
overlap between both phases. The function is independent of the
method used to calculate chill and forcing heat, so that chill can
be supplied as chill hours, chill units or chill portions
(recommended, particularly for warm climates or in climate change studies),
forcing heat can be supplied either as GDD or GDH. The function allows
predicting several stages (or the same for different cultivars), by supplying
a dataframe in which each rows contains chill and heat requirements for a
phenological stage.
}
\examples{

\dontrun{

Predicted_Dates <- phenology_sequential(Weather,Golden_reqs)

}
}
\author{
Carlos Miranda
}
