% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_thermal_time.R
\name{phenology_thermal_time}
\alias{phenology_thermal_time}
\title{Prediction of phenological stages using a thermal time model}
\usage{
phenology_thermal_time(GDH_day, Reqs)
}
\arguments{
\item{GDH_day}{a dataframe with daily forcing accumulation
must contain the columns Year, Month, Day, DOY, GD.}

\item{Reqs}{a dataframe with starting dates and forcing requirements,
must contain the columns Dreq, Freq.}
}
\value{
data frame with the predicted dates for every season in GDH_day
and stages in Reqs. For each season and phenological stage, the information
provided is the date in which forcing begins (Year and DOY), and the
date in which the phenological stage is met (Year and DOY).
It contains the columns Dreq, Freq, Season, Dreq_Year,
Dreq_DOY, Freq_Year, Freq_DOY.
}
\description{
The function predicts phenological phases for a climate series
from daily chill and heat requirements and forcing data.
The thermal time model used in the fucntion consider consider that
only heat accumulated from a set date to a given sum explain the
date of occurrence of the phenological stage (i.e, it assumes that
dormancy release occurs before that date). The function is independent
of the method used to calculate forcing heat, so that forcing heat can
be supplied either as GDD or GDH. The function allows predicting several
stages (or the same for different cultivars), by supplying
a dataframe in which each row contains the day for starting forcing
and heat requirements for a phenological stage.
}
\examples{

\dontrun{

Predicted_Dates <- phenology_thermal_time(Weather,Golden_reqs)

}
}
\author{
Carlos Miranda
}
