% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollination_weather.R
\name{pollination_weather}
\alias{pollination_weather}
\title{Evaluation of weather conditions for pollination on a daily series}
\usage{
pollination_weather(climdata, fendata, lat)
}
\arguments{
\item{climdata}{a dataframe with daily maximum and minimum temperatures,
wind speed and precipitation. Must contain the columns Year, Month, Day,
Tmax, Tmin, WSmed (daily mean wind speed) and Prec (precipitation)}

\item{fendata}{a dataframe with julian day of the beginning (sbloom)
and end (ebloom) of the flowering season. Must contain the columns
Year, sbloom and ebloom in that order.}

\item{lat}{latitude (decimal format) of the site, used to estimate hourly
temperatures.}
}
\value{
a data frame with the columns Year, Sbloom (bloom start, DOY)
, Ebloom (end of bloom, DOY), Bloom_length (in days), Favor (number of
favorable days), Modfavor (number of moderately favorable days) and
Unfavor (number of unfavorable days).
}
\description{
This function estimates the number of days with conditions
favorable, unfavorable and moderately favorable for insect
pollination of fruit trees during the flowering period using
daily weather data.
}
\details{
Days are classified considering the classification proposed by
Williams and Sims (1977), by accounthing the number of favorable
hours for pollination within a day. One hour is considered favorable
if the temperature is above 12.5 C, the speed of the wind below
4.5 m/s and no rainfal occurs (Williams and Sims, 1977; Ramirez and
Davenport, 2013). Hourly temperatures are calculated using the function
stack_hourly_temps from chillR package (Luedeling et al, 2013), hourly
wind speeds from daily means using a modified version of the formula
proposed by (Guo et al, 2016). No hourly downscaling of rainfall is
performed, the function allow daily rainfall below 2.0 mm when
estimating if a day is favorable for pollination or not.
}
\examples{

\dontrun{

#select the appropiate columns from a larger dataset with date information
#in Year, Month, Day format, include date and DOY information and estimate
#the number favorable days on each year in the series

Weather <- Tempdata \%>\%
   select(Year, Month, Day, Tmax, Tmin, WSmed, Prec) \%>\%
   mutate(Date=make_date(Year,Month,Day),DOY=yday(Date))
Pollin_eval <- pollination_weather(Weather,Bloom_D, 41.5)


}
}
\references{
Guo Z, Chang C, Wang R, 2016. A novel method to downscale daily wind
statistics to hourly wind data for wind erosion modelling. In: Bian F.,
Xie Y. (eds) Geo-Informatics in Resource Management and Sustainable
Ecosystem. GRMSE 2015. Communications in Computer and Information Science,
vol 569. Springer, Berlin, Heidelberg

Luedeling E, Kunz A and Blanke M, 2013. Identification of chilling and heat
requirements of cherrytrees - a statistical approach.
International Journal of Biometeorology 57,679-689.

Ramirez F and Davenport TL, 2013. Apple pollination: A review. Scientia
Horticulturae 162, 188-203.

Williams RR, Sims FP, 1977. The importance of weather and variability
in flowering time when deciding pollination schemes for Cox's Orange
Pippin. Experimental Horticulture 29, 15-26.
}
\author{
Carlos Miranda
}
