% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_times.R
\name{solar_times}
\alias{solar_times}
\title{Estimation of the sunrise and sunset hour}
\usage{
solar_times(latitude, DOY)
}
\arguments{
\item{latitude}{numeric value specifying the geographic latitude
(in decimal degrees) of the location of interest}

\item{DOY}{numeric (usually integer) value or vector specifying the
day of the year for which calculations should be done.}
}
\value{
list with Sunrise, Sunset and Daylength.
}
\description{
This function estimates the sunrise and sunset hour
for a location, characterized by latitude, and the
day of the year (DOY). The function uses the equations by
Spencer (1971) and Almorox et al. (2005).
}
\examples{

\dontrun{

#create a vector with 365 days in sequence and calculate sunrise and
#sunset hours for that year in a site placed a 45.5?? N

Days <- seq(1:365)
Sunrise_Sunset <- solar_times(41.5,Days)

}
}
\references{
Spencer JW, 1971. Fourier series representation of the position of the Sun.
Search 2(5), 172.

Almorox J, Hontoria C and Benito M, 2005. Statistical validation of
daylength definitions for estimation of global solar radiation in Toledo,
Spain. Energy Conversion and Management 46(9-10), 1465-1471)
}
\author{
Carlos Miranda
}
