% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spring_frost.R
\name{spring_frost}
\alias{spring_frost}
\title{Calculates the risk of spring frosts for a climate series}
\usage{
spring_frost(tempdata, fendata, tcrit, lastday = 181)
}
\arguments{
\item{tempdata}{a dataframe with daily minimum temperatures for each
year in a series. Must contain the columns Year, julian day of year (DOY) and
the minimum daily temperature (Tmin).}

\item{fendata}{a dataframe with julian day of occurrence of the phenological
stages as produced by phenology_thermal_time or phenology_sequential functions.
Must contain the columns Year, Chill_comp and Pheno_date.}

\item{tcrit}{a vector with critical temperatures for all the phenological stages
indicated in fendata AND Chill_comp date.}

\item{lastday}{the last day (day of the year) to evaluate. By default, 
lastday = 181 (June 30th).}
}
\value{
a list with two data frames. The df Days_frost has the columns Year, DOY, 
Tmin, Tcrit and Day_Frost (indicates a day of frost with a 1 if Tmin<=Tcrit). 
The df Total_frosts indicates the total number of frost days each year in the
series.
}
\description{
The function evaluates the number of early and spring frosts
on each season within a climate data series. It compares minimum
daily temperature (Tmin) with the critical temperature (Tcrit).
Daily critical temperatures are linearly interpolated from a
user-provided dataframe with the day of occurrence of the stages on
each season and a vector of critical temperatures for each stage.
}
\details{
The function checks each day if Tmin is below Tcrit and,
if so, that day is considered as a frost day (Fday). The last day
evaluated each year is user defined, by default set at DOY 181, 
to avoid computing autumn and winter frosts.
The function currently works only with phenological dates occuring
within the same year.
}
\examples{

\dontrun{

#select the appropiate columns from a larger dataset with date information
#in Year, Month, Day format, include date and DOY information and estimate
#the number of frost days for the series

Tmins <- Tempdata \%>\%
   select(Year, Month, Day, Tmin) \%>\%
   mutate(Date=make_date(Year,Month,Day),DOY=yday(Date))
Tcrit <- c(-6.6,-4.8,-4,-1.8,-1.4,-2.3)
Frisk <-spring_frost(Tmins,Tcrit,Phen_data)

}
}
\author{
Carlos Miranda
}
