\name{Tudela_DW}
\alias{Tudela_DW}
\docType{data}
\title{
Daily weather data from Tudela, Spain
}
\description{
Daily weather data (2000-2010) from Tudela, Spain, recorded at the Tudela-Montes de Cierzo
automatic weather station by Gobierno de Navarra.
}
\format{
  A data frame with 4018 observations on the following 11 variables.
  \describe{
    \item{\code{Year}}{a numeric vector, the observation year}
    \item{\code{Month}}{a numeric vector, the observation month}
    \item{\code{Day}}{a numeric vector, the observation day}
    \item{\code{Tmax}}{a numeric vector, daily maximum temperature in Celsius}
    \item{\code{Tmin}}{a numeric vector, daily minimum temperature in Celsius}
    \item{\code{RHmax}}{a numeric vector, daily maximum relative humidity in \%}
    \item{\code{RHmin}}{a numeric vector, daily minimum relative humidity in \%}
    \item{\code{Prec}}{a numeric vector, daily rainfall in mm}
    \item{\code{u2med}}{a numeric vector, daily mean wind speed in m s-1}
    \item{\code{u2max}}{a numeric vector, daily maximum wind speed in m s-1}
    \item{\code{Rad}}{a numeric vector, daily solar radiation in MJ m-2 day-1}
  }
}
\source{
http://meteo.navarra.es/estaciones/descargardatos_estacion.cfm?IDEstacion=36
}

