% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDH_asymcur.R
\name{GDH_asymcur}
\alias{GDH_asymcur}
\title{Calculates growing degree hours (GDH) using ASYMCUR method}
\usage{
GDH_asymcur(Hourdata, Tb, Topt, Tcrit)
}
\arguments{
\item{Hourdata}{a dataframe of hourly temperatures. This data frame
must have a column for Year, Month, Day, DOY (day of year), Hour,
Temp (hourly temperature) and Chill (Chill accumulated).}

\item{Tb}{the base temperatures to calculate GDH}

\item{Topt}{the optimal temperatures to calculate GDH}

\item{Tcrit}{the critical temperature}
}
\value{
data frame with daily data. It contains the columns Date,
Year, Month, Day, DOY (day of the year), Chill, and GDH
}
\description{
The function calculates the daily heat unit accumulation (GDH)
from hourly temperature data, using the ASYMCUR model
proposed by Anderson et al (1986). The model is a refinement
of the linear model proposed by Anderson and Seeley (1992) defined
by a base, optimum and critical temperature. Heat accumulation
begins when temperatures are above a minimum (base temperature,
Tb), and growth increases with temperature up to a
point (optimum temperature, Topt) at which there is no longer
an increase. The critical temperature (Tcrit) is the maximum
temperature at which growth will continue. The difference of
ASYMCUR model with the linear is that the former uses an
assymmetric curvilinear relationship to model GDH accumulation.
The function allows the user to define Tb, Topt and Tcrit.
}
\examples{

\dontrun{

GDH <- GDH_asymcur(Weather,4.5,25,36)

}

}
\references{
Anderson JL, Richardson EA and Kesner CD, 1986. Validation of chill
unit and flower bud phenology models for 'Montmorency' sour cherry.
Acta Horticulturae 184, 71-75.
Anderson JL and Seeley SD, 1992. Modelling strategy in pomology:
Development of the Utah models. Acta Horticulturae 313, 297-306.
}
\author{
Carlos Miranda, \email{carlos.miranda@unavarra.es}
}
