% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDH_linear.R
\name{GDH_linear}
\alias{GDH_linear}
\title{Calculates growing degree hours using a linear method}
\usage{
GDH_linear(Hourdata, Tb, Topt, Tcrit)
}
\arguments{
\item{Hourdata}{a dataframe of hourly temperatures. This data frame
must have a column for Year, Month, Day, DOY (day of year),Hour,
Temp (hourly temperature) and Chill (Chill accumulated).}

\item{Tb}{the base temperatures to calculate GDH}

\item{Topt}{the optimal temperatures to calculate GDH}

\item{Tcrit}{the critical temperature}
}
\value{
data frame with daily data. It contains the columns Date,
Year, Month, Day, DOY (day of the year), Chill, and GDH
}
\description{
The function calculates the daily heat unit accumulation (GDH)
from hourly temperature data, using the linear model
proposed by Anderson and Seeley (1992). The model is defined
by a base, optimum and critical temperature. Heat accumulation
begins when temperatures are above a minimum (base temperature,
Tb), and growth increases linearly with temperature up to a
point (optimum temperature, Topt) at which there is no longer
an increase. The critical temperature (Tcrit) is the maximum
temperature at which growth will continue. The function allows
the user to define Tb, Topt and Tcrit.
}
\examples{

\dontrun{

GDH <- GDH_linear(Weather,4.5,25,36)

}
}
\references{
Anderson JL and Seeley SD, 1992. Modelling strategy in pomology:
Development of the Utah models. Acta Horticulturae 313, 297-306.
}
\author{
Carlos Miranda
}
